requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.11.2
azure-core==1.29.5
azure-storage-blob==12.19.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
distlib==0.3.4
filelock==3.7.1
idna==3.4
isodate==0.6.1
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycparser==2.21
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
typing_extensions==4.8.0
urllib3==2.1.0